% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mm.R
\name{calculate_mm}
\alias{calculate_mm}
\title{Calculate marginal means given a data set}
\usage{
calculate_mm(.data)
}
\arguments{
\item{.data}{A conjoint data set.}
}
\value{
A summarized data set of marginal means for each attribute-level
}
\description{
Calculate marginal means given a data set
}
