% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tau.R
\name{calculate_tau}
\alias{calculate_tau}
\title{Calculate intra-respondent reliability (IRR) from a conjoint data set. This is an internal function.}
\usage{
calculate_tau(.data)
}
\arguments{
\item{.data}{A conjoint data set.}
}
\value{
A scalar indicating IRR.
}
\description{
Calculate intra-respondent reliability (IRR) from a conjoint data set. This is an internal function.
}
