% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_distribution.R
\name{check_distribution}
\alias{check_distribution}
\title{Check the distribution of levels for each attribute}
\usage{
check_distribution(.data)
}
\arguments{
\item{.data}{A data frame specifically for conjiont analysis (use reshape_conjiont())}
}
\value{
A data frame
}
\description{
After generating test responses, it is worth checking the distribution of levels for each attribute using this function.
}
