% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData3.R
\docType{data}
\name{exampleData3}
\alias{exampleData3}
\title{Projoint Example Data Set 3: Building Conjoint with No Repeated Task}
\format{
A data frame with 400 rows and 184 columns:
Contains survey responses including demographic information, outcome choices, 
and conjoint attribute values identified by K-*-* variable names.
}
\usage{
data(exampleData3)
}
\description{
A cleaned Qualtrics output of a conjoint study that compares two potential
new building developments. There are only 8 standard tasks; no repeated tasks are included.
}
\keyword{datasets}
