% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_data.R
\name{organize_data}
\alias{organize_data}
\title{Organize data before estimation}
\usage{
organize_data(
  .dataframe,
  .attribute,
  .level,
  .structure = "choice_level",
  .remove_ties = TRUE,
  .repeated_task = TRUE
)
}
\arguments{
\item{.dataframe}{The data frame -- the second element of the list generated by reshape_conjoint()}

\item{.attribute}{A character identifying the attribute of interest}

\item{.level}{A character vector identifying the levels of interest. Its length should be 1 for profile-level analysis and 2 for choice-level analysis}

\item{.structure}{either "choice_level" or "profile_level"}

\item{.remove_ties}{TRUE if you want to remove ties for the attribute of interest (in profile-level analysis)}

\item{.repeated_task}{TRUE if a repeated task is used to calculate IRR (recommended)}
}
\value{
A data frame
}
\description{
This function ...
}
