% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pj.R
\name{pj}
\alias{pj}
\title{Calculates the intra-respondent reliability (IRR), average marginal component effect (AMCE), marginal mean (MM), and correcte AMCEs and MMs for a given conjoint data set.}
\usage{
pj(.data, .tau = NULL)
}
\arguments{
\item{.data}{A conjoint data set.}

\item{.tau}{An optional numeric argument allowing the researcher to specify their own value of \texttt{tau}}
}
\value{
A summarized data set describing the results of a conjoint analysis, corrected for IRR measurement error.
}
\description{
Calculates the intra-respondent reliability (IRR), average marginal component effect (AMCE), marginal mean (MM), and correcte AMCEs and MMs for a given conjoint data set.
}
\examples{
library(projoint)

data("exampleData1")
head(exampleData1)
reshaped_data = reshape_conjoint(exampleData1, .idvar = "ResponseId",
                                 .outcomes = c("Q4.1", "Q5.1", "Q6.1", "Q7.1", "Q8.1","Q9.1"),
                                 .alphabet ="A")
results = pj(reshaped_data)

library(projoint)

data("exampleData2")
head(exampleData2)
reshaped_data = reshape_conjoint(exampleData1, .idvar = "ResponseId",
                                 .outcomes = c("Q4.1", "Q5.1", "Q6.1", "Q7.1", "Q8.1","Q9.1"),
                                 .alphabet ="A")
results = pj(reshaped_data, .tau = 0.75)
}
