% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pj_amce.R
\name{pj_amce}
\alias{pj_amce}
\title{Adjust AMCEs based on our method}
\usage{
pj_amce(data, formula, id = ~0, n_boot = 100, tau = 0.25, ...)
}
\arguments{
\item{data}{(Description from the cregg package's amce function) A data frame containing variables specified in \code{formula}. All RHS variables should be factors; the base level for each will be used in estimation and its reported AMCE will be NA (for printing). Optionally, this can instead be an object of class \dQuote{survey.design} returned by \code{\link[survey]{svydesign}}.}

\item{formula}{(Description from the cregg package's amce function) A formula specifying an AMCE model to be estimated. All variables should be factors; all levels across features should be unique. Two-way constraints can be specified with an asterisk (*) between RHS features. The specific constrained level pairs within these features are then detected automatically. Higher-order constraints are not allowed.}

\item{n_boot}{The number of bootstrapped samples}

\item{tau}{The estimated/assumed swapping error}
}
\value{
A data frame of class \dQuote{cj_amce}
}
\description{
Adjust AMCEs based on the estimate/assumed measurement error and return a tidy data frame of results
}
