% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pj_bs.R
\name{pj_bs}
\alias{pj_bs}
\title{Calculates the intra-respondent reliability (IRR), average marginal component effect (AMCE), marginal mean (MM), and correcte AMCEs and MMs for a given conjoint data set, as well as bootstrapped confidence intervals.}
\usage{
pj_bs(.data, .id = "id", .n_boot = 500, .diff = NULL)
}
\arguments{
\item{.data}{A conjoint data set.}
}
\value{
A summarized data set describing the results of a conjoint analysis, corrected for IRR measurement error.
}
\description{
Calculates the intra-respondent reliability (IRR), average marginal component effect (AMCE), marginal mean (MM), and correcte AMCEs and MMs for a given conjoint data set, as well as bootstrapped confidence intervals.
}
