% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.projoint_results.R
\name{plot.projoint_results}
\alias{plot.projoint_results}
\title{Plot all MMs or AMCEs}
\usage{
\method{plot}{projoint_results}(
  x,
  .estimand = "mm",
  .estimates = "corrected",
  .by_var = FALSE,
  .base_size = 12,
  .base_family = "",
  ...
)
}
\arguments{
\item{x}{A \code{\link{projoint_results}} object}

\item{.estimand}{Either \code{"mm"} for marginal mean (default) or \code{"amce"} for average marginal component effect}

\item{.estimates}{The estimates to be plotted, either \code{"corrected"} (default), \code{"uncorrected"}, or \code{"both"}}

\item{.by_var}{\code{TRUE} to plot the difference in estimates between the two subgroups, \code{FALSE} (default) otherwise}

\item{.base_size}{base font size, given in pts.}

\item{.base_family}{base font family}

\item{...}{Additional optional arguments}
}
\value{
A \code{ggplot} object
}
\description{
This method produces MM or AMCE plots given a \code{\link{projoint_results}} object, the output from the \code{\link{projoint}} function.
}
\examples{

library(projoint)

data("exampleData1")
head(exampleData1)

outcomes <- paste0("choice", seq(from = 1, to = 8, by = 1))
outcomes <- c(outcomes, "choice1_repeated_flipped")

reshaped_data <- reshape_projoint(
  .dataframe = exampleData1, 
  .idvar = "ResponseId", 
  .outcomes = outcomes,
  .outcomes_ids = c("A", "B"),
  .alphabet = "K", 
  .repeated = TRUE,
  .flipped = TRUE)

projoint_output <- projoint(reshaped_data)
plot(projoint_output)
}
