% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_projoint.R
\name{plot_projoint}
\alias{plot_projoint}
\title{Plot all MMs or AMCEs}
\usage{
plot_projoint(
  .data,
  .estimand = "mm",
  .estimates = "corrected",
  .base_size = 12,
  .base_family = ""
)
}
\arguments{
\item{.data}{A `projoint_results_mm` or `projoint_results_amce` object}

\item{.estimand}{Either "mm" for marginal mean (default) or "amce" for average marginal component effect}

\item{.estimates}{The estimates to be plotted, either "corrected" (default), "uncorrected", or "both"}

\item{.base_size}{base font size, given in pts.}

\item{.base_family}{base font family}
}
\value{
A ggplot object
}
\description{
This function produces MM or AMCE plots given an output from the `projoint` function.
}
\examples{

library(projoint)

data("exampleData1")
head(exampleData1)

 reshaped_data <-  reshape_projoint(
  .dataframe = exampleData1, 
  .idvar = "ResponseId", 
  .outcomes = c(paste0("choice", 1:8), "choice1_repeated_flipped"),
  .outcomes_ids = c("Community A", "Community B"),
  .alphabet = "K", 
  .repeated = TRUE,
  .flipped = TRUE)

projoint_output <- projoint(reshaped_data)
projoint_plot(projoint_output)
}
