% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_tau.R
\name{predict_tau}
\alias{predict_tau}
\title{Estimate tau when there is no repeated task.}
\usage{
predict_tau(.dataframe, .idvar = "id", .title = NULL)
}
\arguments{
\item{.dataframe}{The data frame -- the second element of the list generated by reshape_conjoint()}

\item{.idvar}{A character identifying the column name containing respondent IDs}

\item{.title}{The title of a figure}
}
\value{
A list (a vector of predicted value; ggplot object)
}
\description{
This function ...
}
\examples{
library(projoint)
library(stringr)

data("exampleData1")
head(exampleData1)

outcomes <- str_c("choice", seq(from = 1, to = 8, by = 1))
outcomes <- c(outcomes, "choice1_repeated_flipped")
reshaped_data = reshaped_data <- reshape_conjoint(
  .dataframe = exampleData1, 
  .idvar = "ResponseId", 
  .outcomes = outcomes,
  .outcomes_ids = c("A", "B"),
  .alphabet = "K", 
  .repeated = TRUE,
  .flipped = TRUE)
  
tau <- predict_tau(reshaped_data[[2]])
tau
}
