% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.projoint_data.R
\name{print.projoint_data}
\alias{print.projoint_data}
\title{Print a projoint_data object}
\usage{
\method{print}{projoint_data}(x, ...)
}
\arguments{
\item{x}{A `projoint_data` object.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Custom print method for objects of class `projoint_data`.
}
\keyword{internal}
