% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projoint_data.R
\docType{class}
\name{projoint_data-class}
\alias{projoint_data-class}
\alias{projoint_data}
\title{Class generator for \code{\link{projoint_data}}}
\description{
Class generator for \code{\link{projoint_data}}
}
\keyword{internal}
