% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projoint_data.R
\name{projoint_data}
\alias{projoint_data}
\title{Create a projoint_data Object}
\usage{
projoint_data(labels, data)
}
\arguments{
\item{labels}{A data frame of conjoint attribute and level labels, with attribute and level IDs.}

\item{data}{A data frame containing the reshaped conjoint survey responses.}
}
\value{
An object of class \code{projoint_data}.
}
\description{
This function creates a \code{projoint_data} S3 object containing conjoint survey labels and response data.
It is used internally by \code{\link{reshape_projoint}} and \code{\link{make_projoint_data}}.
}
\keyword{internal}
