% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projoint_irr.R
\docType{class}
\name{projoint_irr-class}
\alias{projoint_irr-class}
\alias{projoint_irr}
\title{Class generator for `projoint_irr`}
\arguments{
\item{slots}{Takes two slots: a scalar for irr, and a ggplot object for figure}
}
\description{
Class generator for `projoint_irr`
}
