% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projoint_qoi.R
\docType{class}
\name{projoint_qoi-class}
\alias{projoint_qoi-class}
\alias{projoint_qoi}
\title{Class generator for `projoint_qoi`}
\arguments{
\item{slots}{Takes two slots: a scalar indicating the attribute of interest, and a scalar or vector indicating the levels of interest}
}
\description{
Class generator for `projoint_qoi`
}
