% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projoint_qoi.R
\docType{class}
\name{projoint_qoi_mm-class}
\alias{projoint_qoi_mm-class}
\alias{projoint_qoi_mm}
\title{Class generator for `projoint_qoi`}
\arguments{
\item{slots}{Takes two or three slots: a scalar indicating the attribute of interest, a scalar or vector indicating the levels of interest, and a scalar or vector indicating the baseline levels of interest (for AMCE)}
}
\description{
Class generator for `projoint_qoi`
}
\keyword{internal}
