% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projoint_results.R
\docType{class}
\name{projoint_results-class}
\alias{projoint_results-class}
\alias{projoint_results}
\title{Class generators for \code{projoint_results}}
\arguments{
\item{slots}{Takes 16 slots. See documentation for \code{\link{projoint}}.}

\item{contains}{Inherits slots from \code{\link{projoint_data}}}
}
\description{
Class generators for \code{projoint_results}
}
\keyword{internal}
