% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projoint_tau.R
\docType{class}
\name{projoint_tau-class}
\alias{projoint_tau-class}
\alias{projoint_tau}
\title{Class generator for \code{projoint_tau} objects}
\arguments{
\item{slots}{Takes two slots: \code{irr} and \code{figure}}

\item{contains}{Inherits slots from \code{\link{projoint_data}}}
}
\description{
Class generator for \code{projoint_tau} objects
}
\keyword{internal}
