% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_labels.R
\name{save_labels}
\alias{save_labels}
\title{Save the attributes and levels, and their order, from a conjoint data set, to a CSV file.}
\usage{
save_labels(.data, .filename)
}
\arguments{
\item{.data}{A \code{\link{projoint_data}} object}

\item{.filename}{The name of a CSV file to be saved}
}
\value{
None
}
\description{
This function saves the labels for attributes and levels, which are generated by \code{\link{reshape_projoint}}, in a CSV file. This is an important step in reorganizing labels.
}
