% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_qoi.R
\name{set_qoi}
\alias{set_qoi}
\title{Set the quantities of interest}
\usage{
set_qoi(
  .structure = "profile_level",
  .estimand = "mm",
  .att_choose,
  .lev_choose,
  .att_notchoose = NULL,
  .lev_notchoose = NULL,
  .att_choose_b = NULL,
  .lev_choose_b = NULL,
  .att_notchoose_b = NULL,
  .lev_notchoose_b = NULL
)
}
\arguments{
\item{.structure}{Either "profile_level" or "choice_level"}

\item{.estimand}{Either "mm" for marginal mean or "amce" for average marginal component effect}

\item{.att_choose}{A character column name identifying the attribute of interest (i.e., for the attribute-level or attribute-levels *chosen*).}

\item{.lev_choose}{A character vector identifying the level or levels of interest (i.e., for the attribute-level or attribute-levels *chosen*). Its length should be 1 for profile-level analysis and 1+ for choice-level analysis}

\item{.att_notchoose}{A character column name identifying the attribute of interest (i.e., for the attribute-level or attribute-levels *not chosen*). This argument should be specified only if the `.structure` argument is "choice-level".}

\item{.lev_notchoose}{A character vector identifying the level or levels of interest (i.e., for the attribute-level or attribute-levels *not chosen*). Its length should be 1 for profile-level analysis and 1+ for choice-level analysis. This argument should be specified only if the `.structure` argument is "choice-level".}

\item{.att_choose_b}{[baseline for AMCE] A character column name identifying the attribute of interest (i.e., for the attribute-level or attribute-levels *chosen*).}

\item{.lev_choose_b}{[baseline for AMCE] A character vector identifying the level or levels of interest (i.e., for the attribute-level or attribute-levels *chosen*). Its length should be 1 for profile-level analysis and 1+ for choice-level analysis}

\item{.att_notchoose_b}{[baseline for AMCE] A character column name identifying the attribute of interest (i.e., for the attribute-level or attribute-levels *not chosen*). This argument should be specified only if the `.structure` argument is "choice-level".}

\item{.lev_notchoose_b}{[baseline for AMCE]  A character vector identifying the level or levels of interest (i.e., for the attribute-level or attribute-levels *not chosen*). Its length should be 1 for profile-level analysis and 1+ for choice-level analysis. This argument should be specified only if the `.structure` argument is "choice-level".}
}
\value{
A `projoint_qoi` object
}
\description{
This function allows users to specify more fine-tuned details of their conjoint analysis. In particular, users can set very specific quantities of interest aside from simple AMCEs or MMs.
}
