% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.projoint_data.R
\name{summary.projoint_data}
\alias{summary.projoint_data}
\title{Summarize a projoint_data object}
\usage{
\method{summary}{projoint_data}(object, ...)
}
\arguments{
\item{object}{A `projoint_data` object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A summary list (invisible).
}
\description{
Custom summary method for objects of class `projoint_data`.
}
\keyword{internal}
