% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.projoint_results.R
\name{summary.projoint_results}
\alias{summary.projoint_results}
\title{Summary method for projoint_results}
\usage{
\method{summary}{projoint_results}(object, ...)
}
\arguments{
\item{object}{An object of class \code{projoint_results}}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for projoint_results
}
\keyword{internal}
