% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tau.R
\name{calculate_tau}
\alias{calculate_tau}
\title{Calculate "tau" (1 - Inter-coder Reliability)}
\usage{
calculate_tau(.data)
}
\arguments{
\item{.data}{A data frame (ashaped for conjoint analysis)}
}
\value{
A numeric
}
\description{
Based on the repeated task, calculate the proportion of tasks with different choices
}
