% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_distribution.R
\name{check_distribution}
\alias{check_distribution}
\title{check the distribution of levels for each attribute}
\usage{
check_distribution(.data)
}
\arguments{
\item{.data}{A data frame (ashaped for conjoint analysis)}
}
\value{
A data-frame
}
\description{
This functino is perhaps useful using test responses. Check the distibution of levels of each attribute.
}
