% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData1.R
\docType{data}
\name{exampleData1}
\alias{exampleData1}
\title{Projoint Example Data Set 1: Building Conjoint with a Repeated, Flipped Task}
\format{
A data frame with 400 rows and 185 columns:
Contains survey responses including demographic information, outcome choices, 
and conjoint attribute values identified by K-*-* variable names.
}
\usage{
data(exampleData1)
}
\description{
A cleaned Qualtrics output of a conjoint study that compares two potential
new building developments. There are 8 standard tasks as well as a repeat
of the first task, used to calculate response instability.
}
\keyword{datasets}
