% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData1_labelled_tibble.R
\name{exampleData1_labelled_tibble}
\alias{exampleData1_labelled_tibble}
\title{Projoint Example Data Set 1: "Labelled Tibble"}
\format{
A data frame with 6,400 rows and 14 columns:
Contains survey responses including demographic information, outcome choices, 
and conjoint attribute values identified by K-*-* variable names.
}
\source{
Qualtrics survey on Prolific; see Clayton et al. replication materials.
}
\description{
An already cleaned data frame (tibble) where each attribute corresponds to a separate column with the proper attribute name. 
The unit of observation is each of two profiles in each task for each respondent.
}
\keyword{datasets}
