% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData3.R
\docType{data}
\name{exampleData3}
\alias{exampleData3}
\title{Projoint Example Data Set 3: Building Conjoint without a Repeated Task}
\format{
## `exampleData3`
A data frame with 469 rows and 132 columns:
\describe{
  \item{ResponseId}{Unique respondent ID}
  \item{gender}{Respondent gender}
  \item{age}{Year}
  \item{choice1, choice2, ..., choice8}{Respondent selections for the initial 8 conjoint tasks}
  \item{K-1-1, K-1-2, ..., K-10-6}{K-X-Y indicates the name of the attribute that is in the Yth position in the table of the Xth task}
  \item{K-1-1-1, K-1-1-2, ..., K-10-2-4}{K-X-Y-Z indicates the value of the attribute that is in the Zth position for the Yth profile (either 1 or 2 for left and right profiles) for the Xth task}
}
}
\source{
Qualtrics and Prolific; see Clayton et al. replication materials.
}
\usage{
exampleData3
}
\description{
A cleaned Qualtrics output of a conjoint study that compares two potential
new building developments. There are 8 standard tasks.
}
\keyword{datasets}
