% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_data.R
\name{organize_data}
\alias{organize_data}
\title{Organize data before estimation}
\usage{
organize_data(
  .dataframe,
  .structure,
  .estimand,
  .remove_ties,
  .att_choose,
  .lev_choose,
  .att_notchoose,
  .lev_notchoose
)
}
\arguments{
\item{.dataframe}{A data frame to reorganize -- the second element of the list generated by reshape_conjoint()}

\item{.structure}{either "choice_level" or "profile_level"}

\item{.remove_ties}{TRUE if you want to remove ties for the attribute of interest (in profile-level analysis)}

\item{.att_choose}{A character column name identifying the attribute of interest (i.e., for the attribute-level or attribute-levels *chosen*).}

\item{.lev_choose}{A character vector identifying the level or levels of interest (i.e., for the attribute-level or attribute-levels *chosen*). Its length should be 1 for profile-level analysis and 1+ for choice-level analysis}

\item{.att_notchoose}{A character column name identifying the attribute of interest (i.e., for the attribute-level or attribute-levels *not chosen*). This argument should be specified only if the `.structure` argument is "choice-level".}

\item{.lev_notchoose}{A character vector identifying the level or levels of interest (i.e., for the attribute-level or attribute-levels *not chosen*). Its length should be 1 for profile-level analysis and 1+ for choice-level analysis. This argument should be specified only if the `.structure` argument is "choice-level".}
}
\description{
This function converts a full conjoint data set to a data set structured for analyzing a specific attribute of interest and specific level(s) of interest. This function receives input from reshape_data() and its output feeds into pj_estimate().
}
\keyword{internal}
