% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_data.R
\name{organize_data}
\alias{organize_data}
\title{Organize data before estimation}
\usage{
organize_data(.dataframe, .attribute, .level, .structure, .remove_ties)
}
\arguments{
\item{.dataframe}{A data frame to reorganize -- the second element of the list generated by reshape_conjoint()}

\item{.attribute}{A character column name identifying the attribute of interest}

\item{.level}{A character vector identifying the levels of interest. Its length should be 1 for profile-level analysis and 2 for choice-level analysis}

\item{.structure}{either "choice_level" or "profile_level"}

\item{.remove_ties}{TRUE if you want to remove ties for the attribute of interest (in profile-level analysis)}

\item{.baseline}{A character vector identifying the baseline level for AMCE. Its length should be 1 for profile-level analysis and 2 for choice-level analysis}
}
\description{
This function converts a full conjoint data set to a data set structured for analyzing a specific attribute of interest and specific level(s) of interest.
This function receives input from reshape_data() and its output feeds into pj_estimate().
}
\keyword{internal}
