% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out1_arranged.R
\docType{data}
\name{out1_arranged}
\alias{out1_arranged}
\title{Example Output: Manually Rearranged Labels}
\format{
A projoint_data object.
}
\description{
An example projoint_data object after labels were manually rearranged.
}
\keyword{datasets}
