% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out1_arranged.R
\docType{data}
\name{out1_arranged}
\alias{out1_arranged}
\title{Re-shaped example data set 1 (REORGANIZED)}
\format{
## `out1_arranged`
 A projoint object of class `projoint_data` ready to pass to `projoint()`.
\describe{
  \item{To be written}{To be written}
}
}
\source{
To be written
}
\usage{
out1_arranged
}
\description{
This ...
}
\details{
THIS VERSION IS AFTER MANUAL RE-LABELLING AND REORDERING.
}
\keyword{datasets}
