% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pj.R
\name{projoint}
\alias{projoint}
\title{Adjust AMCEs or MMs based on our method}
\usage{
projoint(
  data,
  formula,
  id = ~0,
  estimand = c("amce", "mm"),
  n_boot = 100,
  tau = 0.25,
  ...
)
}
\arguments{
\item{data}{(Description from the cregg package's amce function) A data frame containing variables specified in \code{formula}. All RHS variables should be factors; the base level for each will be used in estimation and its reported AMCE will be NA (for printing). Optionally, this can instead be an object of class \dQuote{survey.design} returned by \code{\link[survey]{svydesign}}.}

\item{formula}{(Description from the cregg package's amce function) A formula specifying an AMCE model to be estimated. All variables should be factors; all levels across features should be unique. Two-way constraints can be specified with an asterisk (*) between RHS features. The specific constrained level pairs within these features are then detected automatically. Higher-order constraints are not allowed.}

\item{id}{A formula with a single right-hand-side variable indicating the respondent-level identifier}

\item{estimand}{A character string specifying an estimate type. Current options are average marginal component effects (“amce”) or marginal means (“mm”),}

\item{n_boot}{The number of bootstrapped samples}

\item{tau}{The estimated/assumed swapping error}

\item{...}{Optional arguments to pass to \code{amce()}. For documentation see the \code{cregg} library.}
}
\value{
A data frame of class \dQuote{cj_amce}
}
\description{
Adjust Average Marginal Component Effects (AMCEs) or Marginal Means (MMs) based on the estimate/assumed measurement error and return a tidy data frame of results
}
