% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pj.R
\name{pj}
\alias{pj}
\title{Calculates the intra-respondent reliability (IRR), average marginal component effect (AMCE), marginal mean (MM), and correcte AMCEs and MMs for a given conjoint data set.}
\usage{
pj(.data)
}
\arguments{
\item{.data}{A conjoint data set.}
}
\value{
A summarized data set describing the results of a conjoint analysis, corrected for IRR measurement error.
}
\description{
Calculates the intra-respondent reliability (IRR), average marginal component effect (AMCE), marginal mean (MM), and correcte AMCEs and MMs for a given conjoint data set.
}
