% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pj_diff.R
\name{pj_diff}
\alias{pj_diff}
\title{Calculates the intra-respondent reliability (IRR), average marginal component effect (AMCE), marginal mean (MM), and correcte AMCEs and MMs for a given conjoint data set specified across a given subgroup.}
\usage{
pj_diff(.data, .diff)
}
\arguments{
\item{.data}{A conjoint data set.}

\item{.diff}{A dichotomous variable to calculate difference IRR, AMCEs, and MMs across.}
}
\value{
A summarized data set describing the results of a conjoint analysis subgroup comparison, corrected for IRR measurement error.
}
\description{
Calculates the intra-respondent reliability (IRR), average marginal component effect (AMCE), marginal mean (MM), and correcte AMCEs and MMs for a given conjoint data set specified across a given subgroup.
}
