% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pj_estimate.R
\name{pj_estimate}
\alias{pj_estimate}
\title{Estimate and correct MMs or AMCEs}
\usage{
pj_estimate(
  .data,
  .structure = "profile_level",
  .estimand = "mm",
  .att_choose,
  .lev_choose,
  .att_notchoose = NULL,
  .lev_notchoose = NULL,
  .att_choose_b = NULL,
  .lev_choose_b = NULL,
  .att_notchoose_b = NULL,
  .lev_notchoose_b = NULL,
  .se_method = "analytical",
  .irr = NULL,
  .remove_ties = TRUE,
  .ignore_position = NULL,
  .n_sims = NULL,
  .n_boot = NULL,
  .weights_1 = NULL,
  .clusters_1 = NULL,
  .se_type_1 = "classical",
  .weights_2 = NULL,
  .clusters_2 = NULL,
  .se_type_2 = "classical"
)
}
\arguments{
\item{.data}{A `projoint_data` object}

\item{.structure}{Either "profile_level" (default) or "choice_level"}

\item{.estimand}{Either "mm" for marginal mean or "amce" for average marginal component effect}

\item{.att_choose}{A character column name identifying the attribute of interest (i.e., for the attribute-level or attribute-levels *chosen*).}

\item{.lev_choose}{A character vector identifying the level or levels of interest (i.e., for the attribute-level or attribute-levels *chosen*). Its length should be 1 for profile-level analysis and 1+ for choice-level analysis}

\item{.att_notchoose}{A character column name identifying the attribute of interest (i.e., for the attribute-level or attribute-levels *not chosen*). This argument should be specified only if the `.structure` argument is "choice-level".}

\item{.lev_notchoose}{A character vector identifying the level or levels of interest (i.e., for the attribute-level or attribute-levels *not chosen*). Its length should be 1 for profile-level analysis and 1+ for choice-level analysis. This argument should be specified only if the `.structure` argument is "choice-level".}

\item{.att_choose_b}{[baseline for AMCE] A character column name identifying the attribute of interest (i.e., for the attribute-level or attribute-levels *chosen*).}

\item{.lev_choose_b}{[baseline for AMCE] A character vector identifying the level or levels of interest (i.e., for the attribute-level or attribute-levels *chosen*). Its length should be 1 for profile-level analysis and 1+ for choice-level analysis}

\item{.att_notchoose_b}{[baseline for AMCE] A character column name identifying the attribute of interest (i.e., for the attribute-level or attribute-levels *not chosen*). This argument should be specified only if the `.structure` argument is "choice-level".}

\item{.lev_notchoose_b}{[baseline for AMCE]  A character vector identifying the level or levels of interest (i.e., for the attribute-level or attribute-levels *not chosen*). Its length should be 1 for profile-level analysis and 1+ for choice-level analysis. This argument should be specified only if the `.structure` argument is "choice-level".}

\item{.se_method}{c("analytic", "simulation", "bootstrap") description}

\item{.irr}{NULL (default) if IRR is to be calculated using the repeated task. Otherwise, a numerical value}

\item{.remove_ties}{Logical: should ties be removed before estimation? Defaults to TRUE.}

\item{.ignore_position}{TRUE (default) if you ignore the location of profile (left or right. Relevant only if analyzed at the choice level}

\item{.n_sims}{The number of simulations. Relevant only if .se_method == "simulation"}

\item{.n_boot}{The number of bootstrapped samples. Relevant only if .se_method == "bootstrap"}

\item{.weights_1}{the weight to estimate IRR (see `lm_robust()`): NULL (default)}

\item{.clusters_1}{the clusters to estimate IRR (see `lm_robust()`): NULL (default)}

\item{.se_type_1}{the standard error type to estimate IRR (see `lm_robust()`): "classical" (default)}

\item{.weights_2}{the weight to estimate MM or AMCE (see `lm_robust()`): NULL (default)}

\item{.clusters_2}{the clusters to estimate MM or AMCE (see `lm_robust()`): NULL (default)}

\item{.se_type_2}{the standard error type to estimate MM or AMCE (see `lm_robust()`): "classical" (default)}
}
\value{
A data frame of estimates
}
\description{
This is the internal function used to calculate and correct marginal means or average marginal component effects of a conjoint design.
}
\keyword{internal}
