% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pj_estimate.R
\name{pj_estimate}
\alias{pj_estimate}
\title{Estimate and correct MMs or AMCEs}
\usage{
pj_estimate(
  .list,
  .estimand = "mm",
  .method = "analytical",
  .irr = NULL,
  .n_boot = 1000,
  .n_sims = 10000
)
}
\arguments{
\item{.list}{a list generated by organize_data()}

\item{.estimand}{A character identifying the estimand: "mm" or "amce"}

\item{.method}{A character identifying the method of correcting measurement error bias: "analytical", "simulation", or "bootstrap"}

\item{.irr}{NULL if IRR is calculcate using the repeated task. Otherwise, a numerical value}

\item{.n_boot}{The number of bootstrapped samples}

\item{.n_sims}{The number of simulations}
}
\value{
A data frame
}
\description{
This function ...
}
