% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pj_estimate.R
\name{pj_estimate}
\alias{pj_estimate}
\title{Estimate and correct MMs or AMCEs}
\usage{
pj_estimate(
  .data,
  .attribute,
  .level,
  .structure,
  .estimand,
  .se_method,
  .irr,
  .baseline,
  .remove_ties,
  .ignore_position,
  .n_sims,
  .n_boot
)
}
\arguments{
\item{.data}{A `projoint_data` object}

\item{.attribute}{A character column name identifying the attribute of interest}

\item{.level}{A character vector identifying the levels of interest. Its length should be 1 for profile-level analysis and 2 for choice-level analysis}

\item{.structure}{A character identifying a data structure: "choice_level" (default) or "profile_level". Note: the right element of .level is the outcome of choice if .structure == "choice_level"}

\item{.estimand}{A character identifying the estimand: "mm" (default) or "amce"}

\item{.se_method}{A character identifying the method of correcting measurement error bias: "analytical" (default), "simulation", or "bootstrap" (not recommended)}

\item{.irr}{NULL (default) if IRR is calculate using the repeated task. Otherwise, a numerical value}

\item{.baseline}{A character vector identifying the baseline level. Its length should be 1 for profile-level analysis and 2 for choice-level analysis}

\item{.remove_ties}{TRUE (default) if you want to remove ties for the attribute of interest (in profile-level analysis)}

\item{.ignore_position}{TRUE (default) if you ignore the location of profile (left or right. Relevant only if .structure == "choice_level"}

\item{.n_sims}{The number of simulations (default = 10000). Relevant only if .method == "simulation"}

\item{.n_boot}{The number of bootstrapped samples (default = 1000). Relevant only if .method == "bootstrap"}
}
\value{
A data frame
}
\description{
This function ...
}
