% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pj_estimate.R
\name{pj_estimate}
\alias{pj_estimate}
\title{Estimate and correct MMs or AMCEs}
\usage{
pj_estimate(
  .dataframe,
  .attribute,
  .level,
  .structure = "choice_level",
  .remove_ties = TRUE,
  .repeated_task = TRUE,
  .estimand = "mm",
  .method = "analytical",
  .irr = NULL,
  .ignore_position = TRUE,
  .n_sims = 10000,
  .n_boot = 1000
)
}
\arguments{
\item{.dataframe}{A data frame to reorganize -- the second element of the list generated by reshape_conjoint()}

\item{.attribute}{A character column name identifying the attribute of interest}

\item{.level}{A character vector identifying the levels of interest. Its length should be 1 for profile-level analysis and 2 for choice-level analysis}

\item{.structure}{A character identifying a data structure: "choice_level" (default) or "profile_level". Note: the right element of .level is the outcome of choice if .structure == "choice_level"}

\item{.remove_ties}{TRUE (default) if you want to remove ties for the attribute of interest (in profile-level analysis)}

\item{.repeated_task}{TRUE (default) if a repeated task is used to calculate IRR (recommended)}

\item{.estimand}{A character identifying the estimand: "mm" (default) or "amce"}

\item{.method}{A character identifying the method of correcting measurement error bias: "analytical" (default), "simulation", or "bootstrap" (not recommended)}

\item{.irr}{NULL (default) if IRR is calculate using the repeated task. Otherwise, a numerical value}

\item{.ignore_position}{TRUE (default) if you ignore the location of profile (left or right. Relevant only if .structure == "choice_level"}

\item{.n_sims}{The number of simulations (default = 10000). Relevant only if .method == "simulation"}

\item{.n_boot}{The number of bootstrapped samples (default = 1000). Relevant only if .method == "bootstrap"}
}
\value{
A data frame
}
\description{
This function ...
}
