% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pj_estimate.R
\name{pj_estimate}
\alias{pj_estimate}
\title{Estimate and correct MMs or AMCEs}
\usage{
pj_estimate(
  .data,
  .attribute,
  .level,
  .structure = "choice_level",
  .estimand = "mm",
  .se_method = "analytical",
  .irr = NULL,
  .baseline = NULL,
  .remove_ties = TRUE,
  .ignore_position = NULL,
  .n_sims = NULL,
  .n_boot = NULL,
  .weights_1 = NULL,
  .clusters_1 = NULL,
  .se_type_1 = "classical",
  .weights_2 = NULL,
  .clusters_2 = NULL,
  .se_type_2 = "classical"
)
}
\arguments{
\item{.data}{A `projoint_data` object}

\item{.attribute}{A character column name identifying the attribute of interest}

\item{.level}{A character vector identifying the levels of interest. Its length should be 1 for profile-level analysis and 2 for choice-level analysis}

\item{.structure}{A character identifying a data structure: "choice_level" (default) or "profile_level". Note: the right element of .level is the outcome of choice if .structure == "choice_level"}

\item{.estimand}{A character identifying the estimand: "mm" (default) or "amce"}

\item{.se_method}{A character identifying the method of correcting measurement error bias: "analytical" (default), "simulation", or "bootstrap" (not recommended)}

\item{.irr}{NULL (default) if IRR is calculate using the repeated task. Otherwise, a numerical value}

\item{.baseline}{A character vector identifying the baseline level. Its length should be 1 for profile-level analysis and 2 for choice-level analysis}

\item{.remove_ties}{TRUE (default) if you want to remove ties for the attribute of interest (in profile-level analysis)}

\item{.ignore_position}{TRUE if you ignore the location of profile (left or right). Relevant only if .structure == "choice_level". Defaults to NULL.}

\item{.n_sims}{The number of simulations. Relevant only if .se_method == "simulation"}

\item{.n_boot}{The number of bootstrapped samples. Relevant only if .se_method == "bootstrap"}

\item{.weights_1}{the weight to estimate IRR (see `lm_robust()`): NULL (default)}

\item{.clusters_1}{the clusters to estimate IRR (see `lm_robust()`): NULL (default)}

\item{.se_type_1}{the standard error type to estimate IRR (see `lm_robust()`): "classical" (default)}

\item{.weights_2}{the weight to estimate MM or AMCE (see `lm_robust()`): NULL (default)}

\item{.clusters_2}{the clusters to estimate MM or AMCE (see `lm_robust()`): NULL (default)}

\item{.se_type_2}{the standard error type to estimate MM or AMCE (see `lm_robust()`): "classical" (default)}
}
\value{
A data frame of estimates
}
\description{
This is the internal function used to calculate and correct marginal means or average marginal component effects of a conjoint design.
}
\keyword{internal}
