% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.projoint_results.R
\name{plot.projoint_results}
\alias{plot.projoint_results}
\title{Plot all MMs or AMCEs}
\usage{
\method{plot}{projoint_results}(
  x,
  .estimates = "corrected",
  .by_var = FALSE,
  .labels = NULL,
  .base_size = 12,
  .base_family = "",
  ...
)
}
\arguments{
\item{x}{A \code{\link{projoint_results}} object}

\item{.estimates}{The estimates to be plotted, either \code{"corrected"} (default), \code{"uncorrected"}, or \code{"both"}}

\item{.by_var}{(only if the structure is profile-level) \code{TRUE} to plot the difference in estimates between the two subgroups, \code{FALSE} (default) otherwise}

\item{.labels}{(only if the structure is choice-level) A chaarcter vector for the x-axis labels}

\item{.base_size}{base font size, given in pts.}

\item{.base_family}{base font family}

\item{...}{Additional optional arguments}
}
\value{
A \code{ggplot} object
}
\description{
This method produces MM or AMCE plots given a \code{\link{projoint_results}} object, the output from the \code{\link{projoint}} function.
}
