% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.projoint_tau.R
\name{plot.projoint_tau}
\alias{plot.projoint_tau}
\title{Visualize the results of the extrapolation method for estimating tau.}
\usage{
\method{plot}{projoint_tau}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{projoint_tau}} derived from \code{\link{predict_tau}}}

\item{...}{Optional arguments; currently none accepted}
}
\description{
Visualize the results of the extrapolation method for estimating tau.
}
