% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_projoint.R
\name{plot_projoint}
\alias{plot_projoint}
\title{Plot all MMs or AMCEs}
\usage{
plot_projoint(
  .data,
  .estimand = "mm",
  .estimates = "corrected",
  .base_size = 12,
  .base_family = ""
)
}
\arguments{
\item{.data}{A `projoint_results_mm` or `projoint_results_amce` object}

\item{.estimand}{Either "mm" for marginal mean or "amce" for average marginal component effect}

\item{.estimates}{The estimates to be plotted, either "corrected", "uncorrected", or "both"}

\item{.base_size}{base font size, given in pts.}

\item{.base_family}{base font family}
}
\value{
A ggplot object
}
\description{
This function ...
}
