% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_projoint.R
\name{plot_projoint}
\alias{plot_projoint}
\title{Plot all MMs or AMCEs}
\usage{
plot_projoint(
  .data,
  .estimand = "mm",
  .estimates = "corrected",
  .base_size = 12,
  .base_family = ""
)
}
\arguments{
\item{.data}{A `projoint_results_mm` or `projoint_results_amce` object}

\item{.estimand}{Either "mm" for marginal mean (default) or "amce" for average marginal component effect}

\item{.estimates}{The estimates to be plotted, either "corrected" (default), "uncorrected", or "both"}

\item{.base_size}{base font size, given in pts.}

\item{.base_family}{base font family}
}
\value{
A ggplot object
}
\description{
This function produces MM or AMCE plots given an output from the `projoint` function.
}
\examples{

library(projoint)

data("exampleData1")
head(exampleData1)

outcomes <- paste0("choice", seq(from = 1, to = 8, by = 1))
outcomes <- c(outcomes, "choice1_repeated_flipped")

reshaped_data <- reshape_projoint(
  .dataframe = exampleData1, 
  .idvar = "ResponseId", 
  .outcomes = outcomes,
  .outcomes_ids = c("A", "B"),
  .alphabet = "K", 
  .repeated = TRUE,
  .flipped = TRUE)

projoint_output <- projoint(reshaped_data)
plot_projoint(projoint_output)
}
