% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_projoint_choice_level_mm.R
\name{plot_projoint_choice_level_mm}
\alias{plot_projoint_choice_level_mm}
\title{Plot all "profile-level" MMs or AMCEs}
\usage{
plot_projoint_choice_level_mm(
  x,
  .estimates = "corrected",
  .labels = NULL,
  .base_size = 12,
  .base_family = "",
  ...
)
}
\arguments{
\item{x}{A \code{\link{projoint_results}} object}

\item{.estimates}{The estimates to be plotted, either \code{"corrected"} (default) or \code{"uncorrected"}}

\item{.labels}{A character vector for the x-axis labels}

\item{.base_size}{base font size, given in pts.}

\item{.base_family}{base font family}

\item{...}{Additional optional arguments}
}
\value{
A \code{ggplot} object
}
\description{
This method produces profile-level MM or AMCE plots given a \code{\link{projoint_results}} object, the output from the \code{\link{projoint}} function. The structure must be profile-level to use this function.
}
\keyword{internal}
