% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_tau.R
\name{predict_tau}
\alias{predict_tau}
\title{Predicts \texttt{tau} from a conjoint data set by extrapolation from similar but non-identical tasks.}
\usage{
predict_tau(.data)
}
\arguments{
\item{.data}{A conjoint data set.}
}
\value{
An estimated \texttt{tau} with confidence interval.
}
\description{
Predicts \texttt{tau} from a conjoint data set by extrapolation from similar but non-identical tasks.
}
\examples{
library(projoint)


data("exampleData2")
head(exampleData2)
reshaped_data = reshape_conjoint(exampleData1, .idvar = "ResponseId",
                                 .outcomes = c("Q4.1", "Q5.1", "Q6.1", "Q7.1", "Q8.1","Q9.1"),
                                 .alphabet ="A")
tau_est = predict_tau(reshaped_data)
}
