% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_tau.R
\name{predict_tau}
\alias{predict_tau}
\title{Estimate tau when there is no repeated task.}
\usage{
predict_tau(.data, .title = NULL)
}
\arguments{
\item{.data}{A `projoint_data` object}

\item{.title}{The title of a figure}
}
\value{
A list (a vector of predicted value; ggplot object)
}
\description{
Given an output from `projoint_data`, this function uses the extrapolation method to produce an estimate of intra-coder reliability (IRR).
}
\examples{
library(projoint)
library(stringr)

## Example 1: repeated, flipped task
data("exampleData1")
head(exampleData1)

outcomes <- str_c("choice", seq(from = 1, to = 8, by = 1))
outcomes <- c(outcomes, "choice1_repeated_flipped")
reshaped_data = reshaped_data <- reshape_projoint(
  .dataframe = exampleData1, 
  .idvar = "ResponseId", 
  .outcomes = outcomes,
  .outcomes_ids = c("A", "B"),
  .alphabet = "K", 
  .repeated = TRUE,
  .flipped = TRUE)
  
tau1 <- predict_tau(reshaped_data)
tau1

## Example 2: repeated, unflipped task
data("exampleData2")
head(exampleData2)

outcomes <- str_c("choice", seq(from = 1, to = 8, by = 1))
outcomes <- c(outcomes, "choice1_repeated_notflipped")
reshaped_data = reshaped_data <- reshape_projoint(
  .dataframe = exampleData2, 
  .idvar = "ResponseId", 
  .outcomes = outcomes,
  .outcomes_ids = c("A", "B"),
  .alphabet = "K", 
  .repeated = TRUE,
  .flipped = FALSE)
  
tau2 <- predict_tau(reshaped_data)
tau2

## Example 3: no repeated task
data("exampleData3")
head(exampleData3)

outcomes <- str_c("choice", seq(from = 1, to = 8, by = 1))
outcomes <- c(outcomes)
reshaped_data = reshaped_data <- reshape_projoint(
  .dataframe = exampleData3, 
  .idvar = "ResponseId", 
  .outcomes = outcomes,
  .outcomes_ids = c("A", "B"),
  .alphabet = "K", 
  .repeated = FALSE)
  
tau3 <- predict_tau(reshaped_data)
tau3
}
