% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_tau.R
\name{predict_tau}
\alias{predict_tau}
\title{Estimate tau when there is no repeated task.}
\usage{
predict_tau(.data, .title = NULL)
}
\arguments{
\item{.data}{A \code{\link{projoint_data}} object, outputted from \code{\link{reshape_projoint}}}

\item{.title}{The title of a figure}
}
\value{
A \code{\link{projoint_tau}} object containing the estimate of tau and a figure visualizing the extrapolation method.
}
\description{
Given an output from \code{\link{projoint_data}}, this function uses the extrapolation method to produce an estimate of intra-coder reliability (IRR).
}
