% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_tau.R
\name{predict_tau}
\alias{predict_tau}
\alias{print.projoint_tau}
\alias{summary.projoint_tau}
\alias{plot.projoint_tau}
\title{Estimate tau when there is no repeated task.}
\usage{
predict_tau(.data, .title = NULL)

\method{print}{projoint_tau}(x, ...)

\method{summary}{projoint_tau}(object, ...)

\method{plot}{projoint_tau}(x, ...)
}
\arguments{
\item{.data}{A `projoint_data` object, outputted from `reshape_projoint`}

\item{.title}{The title of a figure}

\item{x}{An object of class \code{\link{projoint_tau}} derived from \code{\link{predict_tau}}}

\item{...}{Optional arguments; currently none accepted}

\item{object}{An object of class \code{\link{projoint_tau}} derived from \code{\link{predict_tau}}}
}
\value{
A list (a vector of predicted value; ggplot object)

A tibble showing IRR as profile similarity changes. The value of \code{predicted} when \code{x=0} is the estimated IRR.
}
\description{
Given an output from `projoint_data`, this function uses the extrapolation method to produce an estimate of intra-coder reliability (IRR).
}
\examples{
library(projoint)
library(stringr)

## Example 1: repeated, flipped task
data("exampleData1")
head(exampleData1)

outcomes <- paste0("choice", seq(from = 1, to = 8, by = 1))
outcomes <- c(outcomes, "choice1_repeated_flipped")
reshaped_data <- reshape_projoint(
  .dataframe = exampleData1, 
  .idvar = "ResponseId", 
  .outcomes = outcomes,
  .outcomes_ids = c("A", "B"),
  .alphabet = "K", 
  .repeated = TRUE,
  .flipped = TRUE)
  
tau1 <- predict_tau(reshaped_data)
tau1

## Example 2: repeated, unflipped task
data("exampleData2")
head(exampleData2)

outcomes <- paste0("choice", seq(from = 1, to = 8, by = 1))
outcomes <- c(outcomes, "choice1_repeated_notflipped")
reshaped_data <- reshape_projoint(
  .dataframe = exampleData2, 
  .idvar = "ResponseId", 
  .outcomes = outcomes,
  .outcomes_ids = c("A", "B"),
  .alphabet = "K", 
  .repeated = TRUE,
  .flipped = FALSE)
  
tau2 <- predict_tau(reshaped_data)
tau2

## Example 3: no repeated task
data("exampleData3")
head(exampleData3)

outcomes <- paste0("choice", seq(from = 1, to = 8, by = 1))
outcomes <- c(outcomes)
reshaped_data <- reshape_projoint(
  .dataframe = exampleData3, 
  .idvar = "ResponseId", 
  .outcomes = outcomes,
  .outcomes_ids = c("A", "B"),
  .alphabet = "K", 
  .repeated = FALSE)
  
tau3 <- predict_tau(reshaped_data)
tau3
}
