% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_tau.R
\name{predict_tau}
\alias{predict_tau}
\title{Estimate tau when there is no repeated task.}
\usage{
predict_tau(.list, .idvar = "id")
}
\arguments{
\item{.list}{A list generated by reshape_conjoint()}

\item{.idvar}{A character identifying the column name containing respondent IDs}
}
\description{
This function ...
}
