% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.projoint_data.R
\name{print.projoint_data}
\alias{print.projoint_data}
\title{Print a projoint_data object}
\usage{
\method{print}{projoint_data}(x, ...)
}
\arguments{
\item{x}{A \code{projoint_data} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
No return value, called for its side effect of printing a summary of
  the \code{projoint_data} object to the console.
}
\description{
Custom print method for objects of class \code{projoint_data}.
}
\examples{
\donttest{
data(exampleData1)
dat <- reshape_projoint(
  exampleData1,
  .outcomes = c("choice1", "choice2")
)
print(dat)
}
}
