% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.projoint_results.R
\name{print.projoint_results}
\alias{print.projoint_results}
\title{Print method for projoint_results}
\usage{
\method{print}{projoint_results}(x, ...)
}
\arguments{
\item{x}{An object of class \code{projoint_results}}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for projoint_results
}
\keyword{internal}
