% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.projoint_tau.R
\name{print.projoint_tau}
\alias{print.projoint_tau}
\title{Print method for projoint_tau objects}
\usage{
\method{print}{projoint_tau}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{projoint_tau}} from \code{\link{predict_tau}}}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print method for projoint_tau objects
}
\keyword{internal}
