% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projoint.R
\name{projoint}
\alias{projoint}
\title{Analyze a conjoint data set and correct for measurement error}
\usage{
projoint(
  .data,
  .structure,
  .irr,
  .se_method = "analytic",
  .ignore_position = TRUE,
  .n_sims = NULL,
  .n_boot = NULL
)
}
\arguments{
\item{.data}{A `projoint.data` object}

\item{.structure}{A `projoint.structure` object. If missing, defaults to producing all MMs and all AMCEs.}

\item{.irr}{NULL (default) if IRR is to be calculated using the repeated task. Otherwise, a numerical value}

\item{.se_method}{c("analytic", "simulation", "bootstrap") description}

\item{.ignore_position}{TRUE (default) if you ignore the location of profile (left or right. Relevant only if analyzed at the choice level}

\item{.n_sims}{The number of simulations. Relevant only if .se_method == "simulation"}

\item{.n_boot}{The number of bootstrapped samples. Relevant only if .se_method == "bootstrap"}
}
\value{
A `projoint.results` object
}
\description{
This function ...
}
