% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projoint.R
\name{projoint}
\alias{projoint}
\alias{print.projoint_results}
\alias{summary.projoint_results}
\title{Analyze a conjoint data set and correct for measurement error}
\usage{
projoint(
  .data,
  .qoi = NULL,
  .by_var = NULL,
  .structure = "profile_level",
  .estimand = "mm",
  .se_method = "analytical",
  .irr = NULL,
  .remove_ties = TRUE,
  .ignore_position = NULL,
  .n_sims = NULL,
  .n_boot = NULL,
  .weights_1 = NULL,
  .clusters_1 = NULL,
  .se_type_1 = "classical",
  .weights_2 = NULL,
  .clusters_2 = NULL,
  .se_type_2 = "classical"
)

\method{print}{projoint_results}(x, ...)

\method{summary}{projoint_results}(object, ...)
}
\arguments{
\item{.data}{A \code{\link{projoint_data}} object}

\item{.qoi}{A \code{\link{projoint_qoi}} object. If \code{NULL}, defaults to producing all MMs and all AMCEs.}

\item{.by_var}{A dichotomous variable (character) used for subgroup analysis}

\item{.structure}{Either \code{"profile_level"} (default) or \code{"choice_level"}. If \code{.qoi} is set, the value of \code{structure} from \code{.qoi} overrides this value.}

\item{.estimand}{Either \code{"mm"} for marginal mean or \code{"amce"} for average marginal component effect. If \code{.qoi} is set, the value of \code{estimand} from \code{.qoi} overrides this value.}

\item{.se_method}{By default, \code{c("analytic", "simulation", "bootstrap")} description}

\item{.irr}{\code{NULL} (default) if IRR is to be calculated using the repeated task. Otherwise, a numerical value}

\item{.remove_ties}{Logical: should ties be removed before estimation? Defaults to \code{TRUE}.}

\item{.ignore_position}{NULL (default) if \code{.structure = "choice_level"}. Set to TRUE if you ignore the position of profile (left or right); FALSE if the relative positioning of profiles matters for analysis. If  \code{.structure = "choice_level"} and this argument is \code{NULL}, it is automatically reset to \code{TRUE}.}

\item{.n_sims}{The number of simulations. Relevant only if \code{.se_method == "simulation"}}

\item{.n_boot}{The number of bootstrapped samples. Relevant only if \code{.se_method == "bootstrap"}}

\item{.weights_1}{the weight to estimate IRR (see \code{\link[estimatr]{lm_robust}}): \code{NULL} (default)}

\item{.clusters_1}{the clusters to estimate IRR (see \code{\link[estimatr]{lm_robust}}): \code{NULL} (default)}

\item{.se_type_1}{the standard error type to estimate IRR (see \code{\link[estimatr]{lm_robust}}): \code{"classical"} (default)}

\item{.weights_2}{the weight to estimate MM or AMCE (see \code{\link[estimatr]{lm_robust}}): \code{NULL} (default)}

\item{.clusters_2}{the clusters to estimate MM or AMCE (see \code{\link[estimatr]{lm_robust}}): \code{NULL} (default)}

\item{.se_type_2}{the standard error type to estimate MM or AMCE (see \code{\link[estimatr]{lm_robust}}): \code{"classical"} (default)}

\item{x}{A \code{\link{projoint_results}} object}

\item{...}{Optional arguments; currently none accepted}

\item{object}{A \code{\link{projoint_results}} object}
}
\value{
A \code{\link{projoint_results}} object
}
\description{
This main function analyzes a conjoint data set and produces measurement error-corrected estimates of either marginal means or average marginal component effects, ready for plotting.
It  accepts a \code{\link{projoint_data}} object, and optionally a \code{\link{projoint_qoi}} object for users who wish to specify more complex quantities of interest.
}
\examples{

library(projoint)

data("exampleData1")
head(exampleData1)

outcomes <- paste0("choice", seq(from = 1, to = 8, by = 1))
outcomes <- c(outcomes, "choice1_repeated_flipped")

reshaped_data <- reshape_projoint(
  .dataframe = exampleData1, 
  .outcomes = outcomes)

projoint(reshaped_data) 
}
