% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projoint.R
\name{projoint}
\alias{projoint}
\title{Analyze a conjoint data set and correct for measurement error}
\usage{
projoint(
  .data,
  .qoi = NULL,
  .structure = "profile_level",
  .estimand = "mm",
  .se_method = "analytical",
  .irr = NULL,
  .remove_ties = TRUE,
  .ignore_position = NULL,
  .n_sims = NULL,
  .n_boot = NULL
)
}
\arguments{
\item{.data}{A `projoint_data` object}

\item{.qoi}{A `projoint_qoi` object. If NULL, defaults to producing all MMs and all AMCEs.}

\item{.structure}{Either "profile_level" or "choice_level"}

\item{.estimand}{Either "mm" for marginal mean or "amce" for average marginal component effect}

\item{.se_method}{c("analytic", "simulation", "bootstrap") description}

\item{.irr}{NULL (default) if IRR is to be calculated using the repeated task. Otherwise, a numerical value}

\item{.remove_ties}{Logical: should ties be removed before estimation? Defaults to TRUE.}

\item{.ignore_position}{TRUE (default) if you ignore the location of profile (left or right. Relevant only if analyzed at the choice level}

\item{.n_sims}{The number of simulations. Relevant only if .se_method == "simulation"}

\item{.n_boot}{The number of bootstrapped samples. Relevant only if .se_method == "bootstrap"}
}
\value{
A `projoint_results` object
}
\description{
This function ...
}
