% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projoint_diff.R
\name{projoint_diff}
\alias{projoint_diff}
\title{Analyze a conjoint data set and correct for measurement error}
\usage{
projoint_diff(
  .data,
  .qoi = NULL,
  .by_var,
  .structure = "profile_level",
  .estimand = "mm",
  .se_method = "analytical",
  .irr = NULL,
  .remove_ties = TRUE,
  .ignore_position = NULL,
  .n_sims = NULL,
  .n_boot = NULL,
  .weights_1 = NULL,
  .clusters_1 = NULL,
  .se_type_1 = "classical",
  .weights_2 = NULL,
  .clusters_2 = NULL,
  .se_type_2 = "classical"
)
}
\arguments{
\item{.data}{A `projoint_data` object}

\item{.qoi}{A `projoint_qoi` object. If NULL, defaults to producing all MMs and all AMCEs.}

\item{.by_var}{A dichotomous variable (character) used for subgroup analysis}

\item{.structure}{Either "profile_level" (default) or "choice_level"}

\item{.estimand}{Either "mm" for marginal mean or "amce" for average marginal component effect}

\item{.se_method}{c("analytic", "simulation", "bootstrap") description}

\item{.irr}{NULL (default) if IRR is to be calculated using the repeated task. Otherwise, a numerical value}

\item{.remove_ties}{Logical: should ties be removed before estimation? Defaults to TRUE.}

\item{.ignore_position}{TRUE (default) if you ignore the location of profile (left or right. Relevant only if analyzed at the choice level}

\item{.n_sims}{The number of simulations. Relevant only if .se_method == "simulation"}

\item{.n_boot}{The number of bootstrapped samples. Relevant only if .se_method == "bootstrap"}

\item{.weights_1}{the weight to estimate IRR (see `lm_robust()`): NULL (default)}

\item{.clusters_1}{the clusters to estimate IRR (see `lm_robust()`): NULL (default)}

\item{.se_type_1}{the standard error type to estimate IRR (see `lm_robust()`): "classical" (default)}

\item{.weights_2}{the weight to estimate MM or AMCE (see `lm_robust()`): NULL (default)}

\item{.clusters_2}{the clusters to estimate MM or AMCE (see `lm_robust()`): NULL (default)}

\item{.se_type_2}{the standard error type to estimate MM or AMCE (see `lm_robust()`): "classical" (default)}
}
\value{
A `projoint_results` object
}
\description{
This is the internal function used to calculate and correct marginal means or average marginal component effects of a conjoint design.
}
\keyword{internal}
