% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projoint_qoi.R
\docType{class}
\name{projoint_qoi-class}
\alias{projoint_qoi-class}
\alias{projoint_qoi}
\title{Class generator for `projoint_qoi`}
\arguments{
\item{slots}{Takes four for MMs and eight for AMCEs}
}
\description{
Class generator for `projoint_qoi`
}
\keyword{internal}
