% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projoint_qoi.R
\name{projoint_qoi}
\alias{projoint_qoi}
\title{Create a projoint_qoi Object}
\usage{
projoint_qoi(...)
}
\arguments{
\item{...}{Named arguments specifying attributes and levels of interest for conjoint estimation.}
}
\value{
An object of class \code{projoint_qoi}.
}
\description{
This function creates a \code{projoint_qoi} S3 object, specifying quantities of interest (QOI) for custom estimations.
It is typically generated by the \code{\link{set_qoi}} function.
}
\keyword{internal}
