% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projoint_results.R
\docType{class}
\name{projoint_results-class}
\alias{projoint_results-class}
\alias{projoint_results}
\title{Class generators for `projoint_results`}
\arguments{
\item{slots}{Takes 16 slots: [to be written]}

\item{contains}{Inherits slots from `projoint_data`}
}
\description{
Class generators for `projoint_results`
}
\keyword{internal}
