% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projoint_results.R
\docType{class}
\name{projoint_results-class}
\alias{projoint_results-class}
\alias{projoint_results}
\title{Class generator for `projoint_results`}
\arguments{
\item{slots}{Takes one slot: `estimate`, a data frame.}

\item{contains}{Inherits slots from `projoint_data`, `projoint_irr`, and, optionally, `projoint_structure`}
}
\description{
Class generator for `projoint_results`
}
