% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projoint_results.R
\docType{class}
\name{projoint_results_mm-class}
\alias{projoint_results_mm-class}
\alias{projoint_results_mm}
\title{Class generator for `projoint_results`}
\arguments{
\item{slots}{Takes one slot: `estimate`, a data frame.}

\item{contains}{Inherits slots from `projoint_data`, `projoint_irr`, and, optionally `projoint_qoi_mm` or `projoint_qoi_amce`}
}
\description{
Class generator for `projoint_results`
}
\keyword{internal}
