% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projoint_results.R
\docType{class}
\name{projoint_results_mm-class}
\alias{projoint_results_mm-class}
\alias{projoint_results_mm}
\title{Class generators for `projoint_results_mm` and `projoint_results_amce`}
\arguments{
\item{slots}{Takes one slot: `estimate`, a data frame.}

\item{contains}{Inherits slots from `projoint_data`, `projoint_irr`, and, optionally `projoint_qoi_mm` or `projoint_qoi_amce`}
}
\description{
Class generators for `projoint_results_mm` and `projoint_results_amce`
}
\keyword{internal}
