% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projoint_tau.R
\docType{class}
\name{projoint_tau-class}
\alias{projoint_tau-class}
\alias{projoint_tau}
\title{Class generator for `projoint_tau`}
\arguments{
\item{slots}{Takes two slots: `irr` and `figure`}

\item{contains}{Inherits slots from `projoint_data`}
}
\description{
Class generator for `projoint_tau`
}
\keyword{internal}
