% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projoint_tau.R
\name{projoint_tau}
\alias{projoint_tau}
\title{Create a projoint_tau Object}
\usage{
projoint_tau(irr, figure)
}
\arguments{
\item{irr}{A tibble containing the extrapolated IRR predictions.}

\item{figure}{A \code{ggplot2} object showing the extrapolation plot.}
}
\value{
An object of class \code{projoint_tau}.
}
\description{
This function creates a \code{projoint_tau} S3 object containing
the estimated reliability (\code{tau}) and the associated diagnostic plot.
}
\keyword{internal}
