% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_labels.R
\name{read_labels}
\alias{read_labels}
\title{Save the original labels for attributes and levels in CSV}
\usage{
read_labels(.data, .filename)
}
\arguments{
\item{.data}{A `projoint_data` object}

\item{.filename}{The name of a REVISED CSV file after manual arrangement}
}
\value{
A projoint object of class `projoint_data` ready to pass to `projoint()`.
}
\description{
This function saves the labels for attributes and levels, which are generated by `reshape_conjoint()`, in a CSV file
}
\examples{
library(projoint)
library(readr)

data("exampleData1")
head(exampleData1)

# Write outcome column names
outcomes <- paste0("choice", seq(from = 1, to = 8, by = 1))
outcomes <- c(outcomes, "choice1_repeated_flipped")

# Reshape the data
reshaped_data <- reshape_projoint(
  .dataframe = exampleData1, 
  .idvar = "ResponseId", 
  .outcomes = outcomes,
  .outcomes_ids = c("A", "B"),
  .alphabet = "K", 
  .repeated = TRUE,
  .flipped = TRUE)
  
# Save a CSV file (Not Run)
# save_labels(reshaped_data, "data-raw/labels_original.csv")

# Read a revised CSV file (Not Run) 
# out1_arranged <- read_labels("data-raw/labels_arranged.csv")
}
