% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Qualtrics.R
\name{read_Qualtrics}
\alias{read_Qualtrics}
\title{Read the Qualtrics data}
\usage{
read_Qualtrics(.file, .folder = "data/")
}
\arguments{
\item{.file}{The name of a Qualtrics data file (in CSV)}

\item{.folder}{The name of a sub-folder relative to the working directory}
}
\value{
A data frame (wide format) with column names
}
\description{
Read the CSV file downloaded from Qualtrics
}
