% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Qualtrics.R
\name{read_Qualtrics}
\alias{read_Qualtrics}
\title{Reads and reformats a Qualtrics csv.}
\usage{
read_Qualtrics(.file)
}
\arguments{
\item{.file}{A filename corresponding to a csv downloaded from Qualtrics.}
}
\value{
A data frame
}
\description{
Reads and reformats a Qualtrics csv.
}
