% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Qualtrics.R
\name{read_Qualtrics}
\alias{read_Qualtrics}
\title{Read and re-format a Qualtrics csv (choice text)}
\usage{
read_Qualtrics(.file)
}
\arguments{
\item{.file}{A file name corresponding to a csv downloaded from Qualtrics.}
}
\value{
A data frame suitable for reshape_conjoint()
}
\description{
Read and re-format a Qualtrics csv (choice text)
}
\examples{
library(projoint)

# Not run:
# dat <- read_Qualtrics("mummolo_nall_replication.csv")
# head(dat)
}
