% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_conjoint.R
\name{reshape_conjoint}
\alias{reshape_conjoint}
\title{Reshape data for conjoint analysis}
\usage{
reshape_conjoint(.data, .idvar, .outcomes, .alphabet)
}
\arguments{
\item{.data}{A data frame}

\item{.idvar}{A variable name for the respondent identifier}

\item{.outcomes}{A character vector of outcome variables}

\item{.alphabet}{The alphabet used for conjoint attribute-levels}
}
\value{
A data frame (long format) ready for conjoint analysis
}
\description{
Reshape the wide-format data to the long-format data
}
