% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_conjoint.R
\name{reshape_conjoint}
\alias{reshape_conjoint}
\title{Reshapes survey response data for conjoint analysis}
\usage{
reshape_conjoint(.data, .idvar, .outcomes, .alphabet = "F", .flipped = TRUE)
}
\arguments{
\item{.data}{A data frame, preferably from \dQuote{read_Qualtrics()}}

\item{.idvar}{A character identifying the column name containing respondent IDs}

\item{.outcomes}{A character vector identifying the column names that contain outcomes}

\item{.alphabet}{The letter indicating conjoint attributes. If using Strezhnev's package (https://github.com/astrezhnev/conjointsdt) in Qualtrics, the default is "F".}

\item{.flipped}{TRUE if the profiles of the repeated task are flipped (recommended)}
}
\value{
A conjoint task-level data frame (in other words, in long format) ready for conjoint analysis. See \dQuote{pj}.
}
\description{
This function takes a data frame, preferably from \dQuote{read_Qualtrics()}, and reshapes it from wide to long such that each row is a distinct conjoint task rather than a respondent.
}
\examples{
library(projoint)

data("exampleData1")
head(exampleData1)
reshaped_data = reshape_conjoint(exampleData1, .idvar = "ResponseId",
                                 .outcomes = c("Q4.1", "Q5.1", "Q6.1", "Q7.1", "Q8.1","Q9.1"),
                                 .alphabet ="A")
}
