% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_conjoint.R
\name{reshape_conjoint}
\alias{reshape_conjoint}
\title{Reshapes survey response data for conjoint analysis}
\usage{
reshape_conjoint(
  .dataframe,
  .idvar,
  .outcomes,
  .outcomes_ids = c("1", "2"),
  .alphabet = "F",
  .repeated = FALSE,
  .flipped = NULL
)
}
\arguments{
\item{.dataframe}{A data frame, preferably from \dQuote{read_Qualtrics()}}

\item{.idvar}{A character identifying the column name containing respondent IDs}

\item{.outcomes}{A character vector identifying the column names that contain outcomes}

\item{.outcomes_ids}{A vector identifying the profiles of the outcomes -- e.g., c("A", "B")}

\item{.alphabet}{The letter indicating conjoint attributes. If using Strezhnev's package (https://github.com/astrezhnev/conjointsdt) in Qualtrics, the default is "F".}

\item{.repeated}{TRUE if there is a repeated task (recommended). The repeated task should be the same as the first task.}

\item{.flipped}{TRUE if the profiles of the repeated task are flipped (recommended)}
}
\value{
A conjoint task-level data frame (in other words, in long format) ready for conjoint analysis.
}
\description{
This function takes a data frame, preferably from \dQuote{read_Qualtrics()}, and reshapes it from wide to long such that each row is a distinct conjoint task rather than a respondent.
}
\examples{
library(projoint)
library(stringr)

data("exampleData1")
head(exampleData1)

outcomes <- str_c("choice", seq(from = 1, to = 8, by = 1))
outcomes <- c(outcomes, "choice1_repeated_flipped")
reshaped_data = reshaped_data <- reshape_conjoint(
  .dataframe = exampleData1, 
  .idvar = "ResponseId", 
  .outcomes = outcomes,
  .outcomes_ids = c("1", "2"),
  .alphabet = "K", 
  .repeated = TRUE,
  .flipped = TRUE)
}
