% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_projoint.R
\name{reshape_projoint}
\alias{reshape_projoint}
\title{Reshapes survey response data for conjoint analysis}
\usage{
reshape_projoint(
  .dataframe,
  .outcomes,
  .outcomes_ids = c("A", "B"),
  .alphabet = "K",
  .idvar = "ResponseId",
  .repeated = TRUE,
  .flipped = TRUE,
  .covariates = NULL,
  .fill = FALSE
)
}
\arguments{
\item{.dataframe}{A data frame, preferably from \code{\link{read_Qualtrics}}}

\item{.outcomes}{A character vector identifying the column names that contain outcomes. If there is a repeated task, it should be the LAST element in this vector.}

\item{.outcomes_ids}{c("A", "B") (default): a vector identifying the possibilities for the outcome variables -- e.g., c("Candidate A", "Candidate B")}

\item{.alphabet}{"K" (default): a letter indicating conjoint attributes. If using Strezhnev's package (\url{https://github.com/astrezhnev/conjointsdt}) in Qualtrics.}

\item{.idvar}{"ResponseId" (default): a character identifying the column name containing respondent}

\item{.repeated}{TRUE (default) if there is a repeated task (recommended). The repeated task should be the same as the first task.}

\item{.flipped}{TRUE (default) if the profiles of the repeated task are flipped (recommended)}

\item{.covariates}{NULL (default): a character vector identifying respondents' covariates used for subgroup analysis}

\item{.fill}{FALSE (default): A logical vector: TRUE if you want to use information about whether a respondent chose the same profile for the repeated task and "fill" (using the `tidyr` package) missing values for the non-repeated tasks, FALSE (otherwise). If the number of respondents is small, if the number of specific profile pairs of your interest is small, and/or if the number of specific respondent subgroups you want to study is small, it is worth changing this option to TRUE. But please note that `.fill = TRUE` is based on an assumption that IRR is independent of information contained in conjoint tables. Although our empirical tests suggest the validity of this assumption, if you are unsure about it, it is better to use the default value (FALSE).}
}
\value{
A projoint object of class `projoint_data` ready to pass to `projoint()`.

A projoint object of class \code{\link{projoint_data}} ready to pass to \code{\link{projoint}}.
}
\description{
This function takes a data frame, preferably from \code{\link{read_Qualtrics}}, and reshapes it from wide to long such that each row is a distinct conjoint task rather than a respondent.
}
\examples{
library(projoint)

data("exampleData1")
head(exampleData1)

# Write outcome column names
outcomes <- paste0("choice", seq(from = 1, to = 8, by = 1))
outcomes <- c(outcomes, "choice1_repeated_flipped")

# Reshape the data
reshaped_data <- reshape_projoint(
  .dataframe = exampleData1, 
  .outcomes = outcomes)
}
