% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_projoint.R
\name{reshape_projoint}
\alias{reshape_projoint}
\title{Reshapes survey response data for conjoint analysis}
\usage{
reshape_projoint(
  .dataframe,
  .idvar,
  .outcomes,
  .outcomes_ids = c("1", "2"),
  .alphabet = "F",
  .repeated = FALSE,
  .flipped = NULL
)
}
\arguments{
\item{.dataframe}{A data frame, preferably from `read_Qualtrics()`}

\item{.idvar}{A character identifying the column name containing respondent IDs}

\item{.outcomes}{A character vector identifying the column names that contain outcomes. If there is a repeated task, it should be the LAST element in this vector.}

\item{.outcomes_ids}{A vector identifying the possibilities for the outcome variables -- e.g., `c("Candidate A", "Candidate B")`}

\item{.alphabet}{The letter indicating conjoint attributes. If using Strezhnev's package (https://github.com/astrezhnev/conjointsdt) in Qualtrics, the default is "F".}

\item{.repeated}{TRUE if there is a repeated task (recommended). The repeated task should be the same as the first task.}

\item{.flipped}{TRUE if the profiles of the repeated task are flipped (recommended)}
}
\value{
A projoint object of class `projoint.data` ready to pass to `projoint()`.
}
\description{
This function takes a data frame, preferably from`read_Qualtrics()`, and reshapes it from wide to long such that each row is a distinct conjoint task rather than a respondent.
}
\examples{
library(projoint)
library(stringr)

data("exampleData1")
head(exampleData1)

# Write outcome column names
outcomes <- str_c("choice", seq(from = 1, to = 8, by = 1))
outcomes <- c(outcomes, "choice1_repeated_flipped")

# Reshape the data
reshaped_data <- reshape_projoint(
  .dataframe = exampleData1, 
  .idvar = "ResponseId", 
  .outcomes = outcomes,
  .outcomes_ids = c("A", "B"),
  .alphabet = "K", 
  .repeated = TRUE,
  .flipped = TRUE)
}
