% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_labels.R
\name{save_labels}
\alias{save_labels}
\title{Save the attributes and levels, and their order, from a conjoint data set, to a CSV file.}
\usage{
save_labels(.data, .filename)
}
\arguments{
\item{.data}{A `projoint_data` object}

\item{.filename}{The name of a CSV file to be saved}
}
\value{
None
}
\description{
This function saves the labels for attributes and levels, which are generated by `reshape_conjoint()`, in a CSV file
}
\examples{
library(projoint)
library(readr)

data("exampleData1")
head(exampleData1)

# Write outcome column names
outcomes <- paste0("choice", seq(from = 1, to = 8, by = 1))
outcomes <- c(outcomes, "choice1_repeated_flipped")

# Reshape the data
reshaped_data <- reshape_projoint(
  .dataframe = exampleData1, 
  .idvar = "ResponseId", 
  .outcomes = outcomes,
  .outcomes_ids = c("A", "B"),
  .alphabet = "K", 
  .repeated = TRUE,
  .flipped = TRUE)

# Not Run: Save a CSV file
# save_labels(reshaped_data, "data-raw/labels_original.csv")
}
