% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_qoi.R
\name{set_qoi}
\alias{set_qoi}
\title{Set the quantities of interest}
\usage{
set_qoi(.structure = "choice_level", .attribute, .level)
}
\arguments{
\item{.structure}{Either "choice_level" or "profile_level}

\item{.attribute}{A character scalar indicating the attribute of interest}

\item{.level}{The level(s) of interest: either a character or a character vector of length 2}
}
\value{
A `projoint_structure` object
}
\description{
This function allows users to specify more fine-tuned details of their conjoint analysis. In particular, users can set very specific quantities of interest aside from simple AMCEs or MMs.
}
