% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_qoi.R
\name{set_qoi}
\alias{set_qoi}
\title{Set the quantities of interest}
\usage{
set_qoi(
  .structure = "choice_level",
  .estimand = "mm",
  .attribute,
  .level,
  .baseline = NULL
)
}
\arguments{
\item{.structure}{Either "profile_level" or "choice_level"}

\item{.estimand}{Either "mm" for marginal mean or "amce" for average marginal component effect}

\item{.attribute}{A character column name identifying the attribute of interest}

\item{.level}{A character vector identifying the levels of interest. Its length should be 1 for profile-level analysis and 2 for choice-level analysis}

\item{.baseline}{A character vector identifying the baseline level for AMCE. Its length should be 1 for profile-level analysis and 2 for choice-level analysis}
}
\value{
A `projoint_qoi` object
}
\description{
This function allows users to specify more fine-tuned details of their conjoint analysis. In particular, users can set very specific quantities of interest aside from simple AMCEs or MMs.
}
