% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.projoint_tau.R
\name{summary.projoint_tau}
\alias{summary.projoint_tau}
\title{Summary method for projoint_tau objects}
\usage{
\method{summary}{projoint_tau}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{projoint_tau}} from \code{\link{predict_tau}}}

\item{...}{Additional arguments (currently unused)}
}
\value{
A tibble showing IRR estimates by the number of differing attributes.
}
\description{
Summary method for projoint_tau objects
}
\keyword{internal}
